refactor : change package structure
