+++ build.gradle
@@ -1,7 +1,7 @@
 plugins {
 	id 'org.springframework.boot' version '2.2.1.RELEASE'
 	id 'io.spring.dependency-management' version '1.0.8.RELEASE'
-	id "org.asciidoctor.convert" version "1.5.9.2"
+	id "org.asciidoctor.convert" version "1.5.10"
 	id 'java'
 }
