+++ src/main/java/com/woowacourse/caffeine/application/service/ShopService.java
@@ -1,7 +1,7 @@
-package com.woowacourse.caffeine.application;
+package com.woowacourse.caffeine.application.service;
 
-import com.woowacourse.caffeine.application.object.ShopCreateRequest;
-import com.woowacourse.caffeine.application.object.ShopResponse;
+import com.woowacourse.caffeine.application.dto.ShopCreateRequest;
+import com.woowacourse.caffeine.application.dto.ShopResponse;
 import com.woowacourse.caffeine.domain.Shop;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
