+++ src/test/java/com/woowacourse/caffeine/application/MenuItemServiceTest.java
@@ -1,5 +1,7 @@
 package com.woowacourse.caffeine.application;
 
+import com.woowacourse.caffeine.application.service.MenuItemService;
+import com.woowacourse.caffeine.application.service.ShopInternalService;
 import com.woowacourse.caffeine.domain.MenuItem;
 import com.woowacourse.caffeine.domain.Shop;
 import com.woowacourse.caffeine.repository.MenuItemRepository;
