+++ src/test/java/com/woowacourse/caffeine/controller/ShopDocumentationTest.java
@@ -1,10 +1,10 @@
 package com.woowacourse.caffeine.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.woowacourse.caffeine.application.MenuItemService;
-import com.woowacourse.caffeine.application.ShopService;
-import com.woowacourse.caffeine.application.object.ShopCreateRequest;
-import com.woowacourse.caffeine.application.object.ShopResponse;
+import com.woowacourse.caffeine.application.service.MenuItemService;
+import com.woowacourse.caffeine.application.service.ShopService;
+import com.woowacourse.caffeine.application.dto.ShopCreateRequest;
+import com.woowacourse.caffeine.application.dto.ShopResponse;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
